pic_labeler/requirements.txt
Flask
google-api-python-client
mysql-connector-python